feat: add error module

+++ server/src/api/routes/operator.py
@@ -1,51 +1,54 @@
 from fastapi import APIRouter, Query
 from database import OperatorController
+from error import ErrorOperatorHandler, STATUS
 from api.models.operator import OperatorBodyModel
-from api.errors.operator import ErrorAPIOperatorHandler
 from api.constant.delete import TypeDelete
-from error import ErrorOperatorHandler
+from api.errors.operator import ERROR_API
 
 router = APIRouter(prefix="/api/v1")
 
 @router.post("/operator/register")
 def get_operators(body: OperatorBodyModel):
     res = OperatorController.create_operator(body.model_dump())
-    if type(res) == ErrorOperatorHandler: 
-        if res.code == 400: raise ErrorAPIOperatorHandler.BAD_REQUEST()
-        elif res.code == 500: raise ErrorAPIOperatorHandler.UNKNOWN()
+    if type(res) == ErrorOperatorHandler: raise ERROR_API(res)
     else: return res
 
 @router.get("/operator")
 def get_operators():
     res = OperatorController.read_operators()
-    if res is None: raise ErrorAPIOperatorHandler.UNKNOWN()
-    return res
+    if type(res) == ErrorOperatorHandler: raise ERROR_API(res)
+    else: return res
 
 @router.get("/operator/all")
 def get_operators():
     res = OperatorController.read_all_operators()
-    if res is None: raise ErrorAPIOperatorHandler.UNKNOWN()
-    return res
+    if type(res) == ErrorOperatorHandler: raise ERROR_API(res)
+    else: return res
 
-@router.get("/operator/username={username}")
-def get_operators(username: str):
-    if not username: raise ErrorAPIOperatorHandler.USERNAME_REQUIRED()
+@router.get("/operator/")
+def get_operators(username: str = Query(None)):
+    if not username: 
+        error = ErrorOperatorHandler(STATUS.ERROR_400_USERNAME_REQUIRED)
+        raise ERROR_API(error)
     res = OperatorController.read_operator(username)
-    if not res and res == []: raise ErrorAPIOperatorHandler.NOT_FOUND()
-    if res is None: raise ErrorAPIOperatorHandler.UNKNOWN()
-    return res
+    if type(res) == ErrorOperatorHandler: raise ERROR_API(res)
+    else: return res
 
 @router.delete("/operator/")
 def delete_operator(username: str = Query(None), delete: TypeDelete = Query(None)):
-    # if not username: raise ErrorAPIOperatorHandler.USERNAME_REQUIRED()
-    # if not delete: raise ErrorAPIOperatorHandler.BAD_REQUEST()
-    # if delete == TypeDelete.HARD.value: 
-    # elif delete == TypeDelete.SOFT.value: 
-    #     # res = OperatorController.delete_operator_soft(username)
-    #     pass
-    # else: raise ErrorAPIOperatorHandler.BAD_REQUEST()
-    res = OperatorController.delete_operator_hard(username)
+    if not username: 
+        error = ErrorOperatorHandler(STATUS.ERROR_400_USERNAME_REQUIRED)
+        raise ERROR_API(error)
+    if not delete: 
+        error = ErrorOperatorHandler(STATUS.ERROR_400_OPTION_REQUIRED)
+        raise ERROR_API(error)
+    if delete == TypeDelete.HARD.value: 
+        res = OperatorController.delete_operator_hard(username)
+    elif delete == TypeDelete.SOFT.value: 
+        res = OperatorController.delete_operator_soft(username)
+    else: 
+        error = ErrorOperatorHandler(STATUS.ERROR_400_DELETE_NOT_VALID)
+        raise ERROR_API(error)
     if type(res) == ErrorOperatorHandler: 
-        print(res.code, res.message)
-        if res.code == 10: raise ErrorAPIOperatorHandler.ERROR_CODE(404, res.message)
+        raise ERROR_API(res)
     else: return res
